Extend the OpenAPI spec for DQT
As part of the Find My TRN work, we need a way to find a person's
details, including their TRN, from a set of search parameters.

The DQT API currently exposes an endpoint that can be used to retrieve a
teacher's details if you already know their TRN. This isn't suitable for
our use case as we are supporting people who don't know their TRN.

I opted to add a new endpoint, `/trn_details`, rather than co-opt
the root `/teachers` resource. I made an assumption that the top-level
resource URI might want to be reserved for listing teachers when
search is not involved.

Also, this endpoint could return records for people that aren't
teachers, so coupling it to a teacher concept seems incorrect.

When helping someone retrieve their TRN, we ask them for an email
address first and attempt to match on that.

There will be another endpoint that a calling service can use to attempt
a TRN match using other parameters. This will be added in a future
change.

An assumption about the matching process is that we can't guarantee that
a single TRN record will be matched using these details. We are of
the belief that a TRN might match multiple people or a person might have
multiple TRN values.

This assumption means we expect the response of the new endpoint to
return an array of trn match objects. It will be up to the calling service
to disambiguate these results.