Overhaul how authentication state is passed around
Previously we would pass around a 'returnUrl' query parameter which
would point back to the authorize endpoint. In addition, we stored user
information in Session (email, name etc.). Storing this information in
two different places is not ideal, especially since Session data could
easily be overwritten by, say, another tab.

This change consolidates all information into Session, but underneath a
unique ID. The unique ID is so we can support multiple authorization
flows for the same user concurrently without the data 'spilling' from
one to another. The ID is passed around using an 'asid' query parameter.
Every new call to /authorize generates a new ID.

In addition, all sign in-related pages are moved into a SignIn folder
and a convention is added that requires the 'asid' query parameter to be
specified. This is to prevent pages from being accessed outside of an
authorization journey.