Add deploy pipeline (#38)
* Add ci rule for automation

Terraform deployment uses ci rule to automate the prompts. This makes it
easy to run terraform from Github actions workflows.

* Add deployment workflow

Deployment workflow will be used to deploy Azure resources and the
docker image to the app service. This commit appends deploy job after
the build job.

* Add site config for Docker image

Docker image can be deployed with terraform as part of app service
deployment. This commit adds a site config block to deploy the docker
image built in previous job.

* Ignore /bin directory

Bin directory is used to download a ruby script for managing Key Vault
secrets. This needs to be ignored on version control.