Amend config structure for API clients to work with app service
We can't have appsetting keys with hyphens in (since environment
variables cannot have hyphens in) but we want our client IDs to be able
to use hyphens. Flip the structure to use an array of clients instead of
an object, that way the client ID is a value not a key.

Also adds config for Stub Find to the app service.