Harden the TRN callback endpoint
The TRN callback was previously a page you could only hit successfully
once for a given authorization journey, since it's the endpoint that
actually creates the user. That's not ideal for a few reasons; it breaks
the back button and should the call to the DQT API fail (and we show an
error) refreshing the page will never work, since we've already
registered the account.

This changes the endpoint so it can be hit successfully multiple times.
We now stash the created user against the journey ID so we know if we've
already created the user or not. Now should the DQT API call fail then
refreshing the page might actually work.