Add blue/green deployment (#59)
* Add swap slot for apps

Get an Identity app deployments in preprod and production should use
app service swap deployment capability. This ensures that apps are
deployed and smoke tested before swapping to production slot. This
commit adds terraform confiuration to deploy staging slot along with the
apps in preprod and production.

* Add environment deployment jobs

Deployment to Azure app service for all environments of Get an Identity
is configured through Terraform deployment files. Preprod and production
jobs have a staging slot that is used for blue/green deployments.

* Modify Makefile

PaaS authentication code is no longer needed for terraform steps as the
deployment is on Azure. Additional IMAGE_TAG is required to pass Github
SHA which is required in Terraform to deploy the correct docker image on
deployments.