Remove code deployment with terraform
Terraform deployment should not deploy to auth server and stage slot
directly. This is noted to be unreliable way of deploying. This commit
removes application_stack argument to terraform resource deployment to
deploy docker image through pipeline. Addional variables are added to
output the terraform values to be used in pipeline.