Deploy to dev using pipeline task (#105)
* add pipeline deploy step

* Remove code deployment with terraform

Terraform deployment should not deploy to auth server and stage slot
directly. This is noted to be unreliable way of deploying. This commit
removes application_stack argument to terraform resource deployment to
deploy docker image through pipeline. Addional variables are added to
output the terraform values to be used in pipeline.

* Update workflow

Actions workflow is included to have url test after code deployment and
dynamics setting of environment variables for terraform and code deploym
ent.