Introduce a new type for producing public claims
Previously we had a single method on AuthenticationState that yielded
claims. This was used for both the internal cookie and also for
generating ID and access tokens. This was a little messy because it
requires creating an AuthenticationState object all the time and it
conflates internal vs public claims.

This change adds UserClaimHelper.GetPublicClaims which has two
overloads; one that takes an AuthenticationState and one which takes a
User.

This will also simplify adding the 'user info' endpoint.