Issue a UserSignedInEvent the first time protected pages are accessed (#219)
The OAuth authorization journey issues a UserSignedInEvent every time a
user completes the authorization flow, even if they are a known user and
we haven't re-authenticated them. This change ensures we're also issuing
an event the first time the user accesses a protected area on the site
(e.g. admin pages) and they're not going through an OAuth
authorization flow.