Ping /status on slot swap
By default the App Service swap slot process pings the application root
and accepts any status code as confirmation the staging app is ready to
swap to production.  The app has implemented health checks so we should
ping the /status endpoint for a 200 response which will provide more
reliable confirmation the app is ready to be swapped to the production
slot.