Fix infinite sign in loop for staff users (#308)
Our SignIn method used to maintain the original IssuedUtc property for
users who were already signed in. That meant that users who were signing
in again more than 20 minutes after the initial sign in were stuck as
the maxage check in the authorize endpoint was always rejecting their
principal as too old and restarting the journey.

This change adds an additional parameter to the SignIn method -
resetIssues - which, when set, will reset the Issued property on the
AuthenticationTicket. We set that to true only when signing in as part
of the sign in journey.