Simplify Claim creation (#431)
Until now we've kept track of user information in potentially up to
three places; AuthenticationState, the DB and the Cookies Principal.
This is becoming more and more complicated and expensive to maintain and
is problematic when we want to add additional public claims without
storing them in the auth cookie.

This change simplifies things; we keep a minimal set of claims in the
auth cookie and always go to the DB to get the full set of 'public'
claims (i.e. those that go into ID and access tokens). The factory
method on AuthenticationState to create a new instance from an
already-signed in user is also changed to take a User object instead of
ClaimsPrincipal.