Refinements to Change official DOB journey (#527)
* Support binding yyyy-MM-dd dates when BindingSource isn't specified

Our yyyy-MM-dd DateOnly model binder previously only ran when
BindingSource was explicitly set to Path or Query. This was to avoid
interfering with the model binder provided by the GFA library for
working with Date input components. The result of this is that we can't
have a single property that can be bound from either the request URL
*or* the body (since BindingSource is `null` in such cases).

This change removes the BindingSource predicate so we will always
attempt to bind using the yyyy-MM-dd format. If there is no value
provided, we fall back to the GFA DateOnly binder.

* Ensure stated values are retained when coming from the Confirm page

Also retains the uploaded file's name so we can show it on the Confirm
page instead of our internally-generated name.

* Add an e2e test covering DQT DOB change